ai-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
PyPDF2==3.0.1
python-multipart==0.0.6
